stproject/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-jet-reboot==1.3.7
django-material-admin==1.8.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.0
google-api-python-client==1.4.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
psutil==5.9.8
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.29
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0
